feat(geography): modernize api and persistence
Move geography API to Integer IDs end-to-end, add paginated list contracts, split repositories into interface plus impl, adopt MapStruct DTO mappers, harden error and security behavior, and align SQL/config defaults with current runtime behavior.

+++ samples/23-geography/src/main/java/com/learning/geography/city/City.java
@@ -1,4 +1,4 @@
 package com.learning.geography.city;
 
-public record City(Long id, Long stateId, String name) {
+public record City(Integer id, Integer stateId, String name) {
 }
