+++ samples/23-geography/src/main/java/com/learning/geography/city/CityNotFoundException.java
@@ -6,7 +6,7 @@ import com.learning.geography.exception.AppException;
 
 class CityNotFoundException extends AppException {
 
-  CityNotFoundException(Long id) {
+  CityNotFoundException(Integer id) {
     super("error.city.not.found", new Object[] { id }, HttpStatus.NOT_FOUND);
   }
 }
