+++ samples/23-geography/src/main/java/com/learning/geography/city/CityResponse.java
@@ -1,4 +1,4 @@
 package com.learning.geography.city;
 
-public record CityResponse(Long id, Long stateId, String name) {
+public record CityResponse(Integer id, Integer stateId, String name) {
 }
