+++ samples/23-geography/src/main/java/com/learning/geography/country/Country.java
@@ -1,4 +1,4 @@
 package com.learning.geography.country;
 
-public record Country(Long id, String name, String isoCode) {
+public record Country(Integer id, String name, String isoCode) {
 }
