+++ samples/23-geography/src/main/java/com/learning/geography/country/CountryResponse.java
@@ -1,4 +1,4 @@
 package com.learning.geography.country;
 
-public record CountryResponse(Long id, String name, String isoCode) {
+public record CountryResponse(Integer id, String name, String isoCode) {
 }
