+++ samples/23-geography/src/main/java/com/learning/geography/state/State.java
@@ -1,4 +1,4 @@
 package com.learning.geography.state;
 
-public record State(Long id, Long countryId, String name, String abbreviation) {
+public record State(Integer id, Integer countryId, String name, String abbreviation) {
 }
